Add GH action to build, test and upload packages
This GH action gets the modified packages, build them and tests them in
every PR. It also run on every push to `main` and in this case it also
upload the packages to MyGet.

If the building or testing or any of the packages fails, the action
stops and fails. If any of the packages fails to upload, the action
fails but it doesn't stop (it tries to upload the rest of the packages
before existing).

The secret `MYGET_TOKEN` is needed for the packages upload.