Use Ana06/get-changed-files@v1.1
Use the new Ana06/get-changed-files@v1.1. It is a fork of
https://github.com/jitterbit/get-changed-files, which supports
`pull_request_target` and allow to filter files using regular
expressions.  The filter feature allows us to not run the `build`,
`test` and `upload` steps if no `.nuspec` files where modified. Before
there were run and nothing was built, tested or uploaded. So the main
difference is that it makes easier to read the log.