Check daily packages
Basic action that runs once a day to test that all the packages are
working. When the first package fails, the action exits.

It runs at a weird time (03:17 UTC) to avoid delays during periods of
high loads of GitHub Actions workflow runs. High load times include the
start of every hour.