Remove non required metadata from nuspec
Remove the nuspec metadata elements that are not strictly necessary to
spare maintenance time. Our packages are expected to be installed
automatically by a script, so providing detailed information is not a
priority. This also solves the "license problem", as there are no
licenses anymore. Remember we are not distributing software:
https://docs.chocolatey.org/en-us/information/legal

Also, simplify descriptions to avoid changes with version updates.

This should be documented in the Wiki.

Reference: https://docs.microsoft.com/en-us/nuget/reference/nuspec

Closes: https://github.com/fireeye/VM-packages/issues/7