Fix script
The script is broken since we moved it into the `utils` directory.
Take the directory name twice to get now the root directory.
Alternatively we could use a command like the following to get the root
of the Git repository:
```
git rev-parse --show-toplevel
```

But I don't think this is very important, as we shouldn't need this
script again.