Update packages to use new '-innerFolder' flag
Also fixed uniextract2 package. It didn't need the shimming logic.
The shimming logic it used only applies if it was being installed to the default Chocolatey directory in %PROGRAMDATA%\chocolatey\lib\uniextract2.
Chocolatey will auto-shim any .exe files it finds in the default install location. So previously we prevented it from auto-shimming all .exes.
However, now that we're installing to %RAW_TOOLS_DIR%, there is no need as we need to explicitly shim with Install-BinFile().